+++ src/enum/TrackSource.ts
@@ -2,5 +2,6 @@ export enum TrackSource {
     OPEN_SKY_NETWORK = 'OPSN',
     VIRTUAL_RADAR_SERVER = 'VRS',
     FACHADEV = 'FachaDev',
+    RADARPLANE = 'RadarPlane',
     DUMP1090 = 'dump1090',
 }