+++ src/utils/EventUtils.ts
@@ -14,7 +14,8 @@ import {Attachment, login} from "masto";
 
 export class EventUtils {
     public static async triggerEvent(event: PlaneEvents, aircraft: Aircraft | AircraftMeta, airline: Airline | Type | Squawk | null = null, data: any = null) {
-        const adsbExchangeLink = 'https://globe.adsbexchange.com/?icao=' + aircraft.icao;
+        // const adsbExchangeLink = 'https://globe.adsbexchange.com/?icao=' + aircraft.icao;
+        const adsbExchangeLink = 'https://radarplane.com/?hex=' + aircraft.icao;
         let photoUrl = await PlaneSpotterUtils.getPhotoUrl(aircraft.icao);
         const notificationSettings = aircraft instanceof Aircraft ? aircraft.notifications : airline!.notifications;
         const notificationName = aircraft instanceof Aircraft ? aircraft.name : airline!.name;
@@ -29,7 +30,7 @@ export class EventUtils {
                         return;
                     }
                     if (notificationSettings.includeScreenshots) {
-                        hasTakeoffScreenshot = await ScreenshotUtils.takeScreenshot(aircraft.icao);
+                        hasTakeoffScreenshot = await ScreenshotUtils.takeRadarPlaneScreenshot(aircraft.icao);
                     }
                     if (notificationSettings.discord.enabled) {
                         let message = `**${notificationName}**${aircraft.callsign ? " flight " + aircraft.callsign : ""} (${aircraft.registration}) took off from **${data.nearestAirport.name}** at <t:${(new Date().getTime() / 1000).toFixed(0)}:t>\n${adsbExchangeLink}`;
@@ -119,7 +120,7 @@ export class EventUtils {
                         return;
                     }
                     if (notificationSettings.includeScreenshots) {
-                        hasLandingScreenshot = await ScreenshotUtils.takeScreenshot(aircraft.icao);
+                        hasLandingScreenshot = await ScreenshotUtils.takeRadarPlaneScreenshot(aircraft.icao);
                     }
                     if (notificationSettings.discord.enabled) {
                         let message = `**${notificationName}**${aircraft.callsign ? " flight " + aircraft.callsign : ""} (${aircraft.registration}) landed at **${data.nearestAirport.name}** at <t:${(new Date().getTime() / 1000).toFixed(0)}:t>\n${adsbExchangeLink}`;
@@ -206,7 +207,7 @@ export class EventUtils {
                 case PlaneEvents.PLANE_EMERGENCY:
                     let hasEmergencyScreenshot = false;
                     if (notificationSettings.includeScreenshots) {
-                        hasEmergencyScreenshot = await ScreenshotUtils.takeScreenshot(aircraft.icao);
+                        hasEmergencyScreenshot = await ScreenshotUtils.takeRadarPlaneScreenshot(aircraft.icao);
                     }
                     if (notificationSettings.discord.enabled) {
                         let message = `**${notificationName}**${aircraft.callsign ? " flight " + aircraft.callsign : ""} is squawking ${data.squawk} ** (${PlaneUtils.getEmergencyType(data.squawk)}) ** at <t:${(new Date().getTime() / 1000).toFixed(0)}:t>\n${adsbExchangeLink}`;